+++ Api/Controllers/UsuarioController.cs
@@ -1,8 +1,14 @@
-﻿using Domain.Commands.v1.Usuarios.CriarSenha;
+﻿using CrossCutting.Configuration.Authorization;
+using Domain.Commands.v1.Usuarios.ListarUsuarios;
+using Domain.Commands.v1.Usuarios.RemoverUsuario;
+using Domain.Commands.v1.Usuarios.AtualizarUsuario;
+using Domain.Commands.v1.Usuarios.BuscarUsuarioPorId;
+using Domain.Commands.v1.Usuarios.CriarSenha;
 using Domain.Commands.v1.Usuarios.CriarUsuario;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Domain.Commands.v1.Usuarios.AlterarStatusUsuario;
 
 namespace Api.Controllers
 {
@@ -37,6 +43,77 @@ namespace Api.Controllers
         {
             var response = await _mediator.Send(command);
             return Created(string.Empty, response);
+        }
+
+        [HttpGet]
+        [Route(template: "BuscarUsuarios")]
+        [Authorize(Policy = PoliticasDeAcesso.Admin)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<IActionResult> BuscarUsuarios()
+        {
+            var query = new ListarUsuariosCommand();
+            var usuarios = await _mediator.Send(query);
+
+            return Ok(usuarios);
+        }
+
+        [HttpGet]
+        [Route(template: "BuscarUsuarioPorId/{id}")]
+        [Authorize(Policy = PoliticasDeAcesso.Admin)]
+        [ProducesResponseType(typeof(BuscarUsuarioPorIdCommandResponse), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<IActionResult> BuscarUsuarioPorId(Guid id)
+        {
+            var query = new BuscarUsuarioPorIdCommand(id);
+            var usuario = await _mediator.Send(query);
+
+            return Ok(usuario);
+        }
+
+        [HttpPut]
+        [Route(template: "AtualizarUsuario/{id}")]
+        [Authorize(Policy = PoliticasDeAcesso.Admin)]
+        [ProducesResponseType(typeof(AtualizarUsuarioCommandResponse), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<IActionResult> AtualizarUsuario([FromBody] AtualizarUsuarioCommand command)
+        {
+            return Ok(await _mediator.Send(command));
+        }
+
+        [HttpPut]
+        [Route(template: "AlterarStatusUsuario")]
+        [Authorize(Policy = PoliticasDeAcesso.Admin)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<IActionResult> AlterarStatusUsuario([FromBody] AlterarStatusUsuarioCommand command)
+        {
+            var result = await _mediator.Send(command);
+            return Ok(result);
+        }
+
+        [HttpDelete]
+        [Route(template: "RemoverUsuario/{id}")]
+        [Authorize(Policy = PoliticasDeAcesso.Admin)]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<IActionResult> RemoverUsuario(Guid id)
+        {
+            var command = new RemoverUsuarioCommand(id);
+            await _mediator.Send(command);
+
+            return NoContent();
         }
     }
 }
