--- Domain/Commands/v1/Adm/AlteraStatusUser/AlteraUserStatusCommandResponse.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Commands.v1.Adm.AlteraStatusUser
-{
-    public class AlteraUserStatusCommandResponse
-    {
-        public bool EstaAtivo { get; set; }
-    }
-}
