+++ Domain/Commands/v1/Usuarios/AlterarStatusUsuario/AlterarStatusUsuarioCommandResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace Domain.Commands.v1.Usuarios.AlterarStatusUsuario
+{
+    public class AlterarStatusUsuarioCommandResponse
+    {
+        public bool EstaAtivo { get; set; }
+    }
+}
