+++ Domain/Commands/v1/Usuarios/BuscarUsuarioPorId/BuscarUsuarioPorIdCommandHandler.cs
@@ -2,7 +2,7 @@
 using Infrastructure.Data.Interfaces.Usuarios;
 using MediatR;
 
-namespace Domain.Commands.v1.Adm.BuscarUsuarioPorId
+namespace Domain.Commands.v1.Usuarios.BuscarUsuarioPorId
 {
     public class BuscarUsuarioPorIdCommandHandler : IRequestHandler<BuscarUsuarioPorIdCommand, BuscarUsuarioPorIdCommandResponse>
     {
