+++ Domain/Commands/v1/Usuarios/BuscarUsuarioPorId/BuscarUsuarioPorIdCommandValidator.cs
@@ -1,10 +1,11 @@
 ﻿using FluentValidation;
 
-namespace Domain.Commands.v1.Adm.BuscarUsuarioPorId
+namespace Domain.Commands.v1.Usuarios.BuscarUsuarioPorId
 {
     public class BuscarUsuarioPorIdCommandValidator : AbstractValidator<BuscarUsuarioPorIdCommand>
     {
-        public BuscarUsuarioPorIdCommandValidator() {
+        public BuscarUsuarioPorIdCommandValidator()
+        {
 
             RuleFor(command => command.Id)
                 .NotEmpty().WithMessage("O ID do usuário não pode estar vazio.")
