+++ Domain/Commands/v1/Usuarios/ListarUsuarios/ListarUsuariosCommand.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace Domain.Commands.v1.Adm.ListarUsuarios
+namespace Domain.Commands.v1.Usuarios.ListarUsuarios
 {
     public class ListarUsuariosCommand : IRequest<IEnumerable<ListarUsuariosCommandResponse>>
     {
