+++ Domain/Commands/v1/Usuarios/ListarUsuarios/ListarUsuariosCommandResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace Domain.Commands.v1.Adm.ListarUsuarios
+﻿namespace Domain.Commands.v1.Usuarios.ListarUsuarios
 {
     public class ListarUsuariosCommandResponse
     {
