+++ Domain/Commands/v1/Usuarios/RemoverUsuario/RemoverUsuarioCommand.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace Domain.Commands.v1.Adm.RemoverUsuario
+namespace Domain.Commands.v1.Usuarios.RemoverUsuario
 {
     public class RemoverUsuarioCommand : IRequest<Unit>
     {
