+++ Domain/Commands/v1/Usuarios/RemoverUsuario/RemoverUsuarioCommandHandler.cs
@@ -1,7 +1,7 @@
 ﻿using Infrastructure.Data.Interfaces.Usuarios;
 using MediatR;
 
-namespace Domain.Commands.v1.Adm.RemoverUsuario
+namespace Domain.Commands.v1.Usuarios.RemoverUsuario
 {
     public class RemoverUsuarioCommandHandler : IRequestHandler<RemoverUsuarioCommand, Unit>
     {
