+++ Domain/Commands/v1/Usuarios/RemoverUsuario/RemoverUsuarioCommandValidator.cs
@@ -1,6 +1,6 @@
 ﻿using FluentValidation;
 
-namespace Domain.Commands.v1.Adm.RemoverUsuario
+namespace Domain.Commands.v1.Usuarios.RemoverUsuario
 {
     public class RemoverUsuarioCommandValidator : AbstractValidator<RemoverUsuarioCommand>
     {
