+++ Domain/MapperProfiles/UsuarioProfile.cs
@@ -1,10 +1,10 @@
 ﻿using AutoMapper;
-using Domain.Commands.v1.Adm.ListarUsuarios;
 using Domain.Commands.v1.Usuarios.CriarUsuario;
 using Infrastructure.Data.Models.Usuarios;
-using Domain.Commands.v1.Adm.AtualizarUsuario;
-using Domain.Commands.v1.Adm.BuscarUsuarioPorId;
-using Domain.Commands.v1.Adm.CadastrarUsuario;
+using Domain.Commands.v1.Usuarios.ListarUsuarios;
+using Domain.Commands.v1.Usuarios.AtualizarUsuario;
+using Domain.Commands.v1.Usuarios.BuscarUsuarioPorId;
+using Domain.Commands.v1.Usuarios.AlterarStatusUsuario;
 
 namespace Domain.MapperProfiles
 {
@@ -12,13 +12,16 @@ namespace Domain.MapperProfiles
     {
         public UsuarioProfile()
         {
-            CreateMap<UsuarioModel, CadastrarUsuarioCommandResponse>();
             CreateMap<UsuarioModel, CriarUsuarioCommandResponse>();
             CreateMap<UsuarioModel, ListarUsuariosCommandResponse>();
             CreateMap<UsuarioModel, AtualizarUsuarioCommandResponse>();
             CreateMap<UsuarioModel, BuscarUsuarioPorIdCommandResponse>();
             CreateMap<CriarUsuarioCommand, UsuarioModel>();
-            CreateMap<CadastrarUsuarioCommand, UsuarioModel>();
+            CreateMap<AtualizarUsuarioCommand, UsuarioModel>();
+            CreateMap<AlterarStatusUsuarioCommand, UsuarioModel>();
+            CreateMap<UsuarioModel, AlterarStatusUsuarioCommandResponse>();
+            CreateMap<Boolean, AlterarStatusUsuarioCommandResponse>()
+                .ForMember(dest => dest.EstaAtivo, opt => opt.MapFrom(src => src));
         }
     }
 }
