+++ Infrastructure/Infrastructure.Data/AppDbContext.cs
@@ -1,6 +1,4 @@
-﻿using Infrastructure.Data.Models;
-using Infrastructure.Data.Models.Adm.AlteraStatusUser;
-using Infrastructure.Data.Models.Jogos;
+﻿using Infrastructure.Data.Models.Jogos;
 using Infrastructure.Data.Models.Promocao;
 using Infrastructure.Data.Models.Usuarios;
 using Microsoft.EntityFrameworkCore;
@@ -22,7 +20,6 @@ namespace Infrastructure.Data
         public DbSet<JogoModel> Jogos { get; set; }
         public DbSet<PromocaoModel> Promocoes { get; set; }
         public DbSet<PromocaoJogoModel> PromocaoJogos { get; set; }
-
         public DbSet<UsuarioModel> Usuarios { get; set; }
     }
 }
