+++ Infrastructure/Infrastructure.Data/Repositories/Usuarios/UsuarioRepository.cs
@@ -27,7 +27,6 @@ namespace Infrastructure.Data.Repositories.Usuarios
 
         public async Task<UsuarioModel> ObterPorIdAsync(Guid id)
         {
-            //TO DO tratar retorno null?
             return await _context.Usuarios.FindAsync(id);
         }
 
@@ -48,6 +47,21 @@ namespace Infrastructure.Data.Repositories.Usuarios
         {
             _context.Usuarios.Remove(usuario);
             await _context.SaveChangesAsync();
+        }
+
+        public async Task<bool> AlterarStatusUsuario(UsuarioModel userModel)
+        {
+            var user = await _context.Usuarios.Where(x => x.Id == userModel.Id).FirstOrDefaultAsync();
+            if (user == null)
+            {
+                throw new KeyNotFoundException("Usuário não encontrado");
+            }
+            if (user.Ativo != userModel.Ativo)
+            {
+                user.Ativo = userModel.Ativo;
+                await _context.SaveChangesAsync();
+            }
+            return user.Ativo;
         }
     }
 }
